reorganize code and update unit tests
